Built working prototype of render option that doesn't require relative positing
the wrap rendermode works, but it has a quirk where if you change the number of columns the order of the pins is not preserved.  This may or maynot be an issue depending on your app.  This issue can be resolved if we find a way to keep track of the order of the pins spearate from their order in columns